fix: change helia-bundle namespace
